[JBPM-8471] Add header and some minimal css
