+++ process-migration-service/frontend/src/css/pim-main.css
@@ -2,3 +2,8 @@
   content:"*";
   color:red;
 }
+
+#pim-content {
+  margin-right: 20px;
+  margin-left: 20px;
+}
