+++ process-migration-service/frontend/src/index.html
@@ -2,6 +2,7 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
+  <title>Process instance migration</title>
 </head>
 
 <body>
