+++ process-migration-service/frontend/src/index.js
@@ -5,11 +5,11 @@ import "patternfly/dist/css/patternfly-additions.min.css";
 import "react-datetime/css/react-datetime.css";
 import "./css/pim-main.css";
 
-import MainPageWithPfTab from "./component/MainPageWithPfTab";
+import MainPage from "./component/MainPage";
 
 document.addEventListener("DOMContentLoaded", function() {
   ReactDOM.render(
-    React.createElement(MainPageWithPfTab),
+    React.createElement(MainPage),
     document.getElementById("mount")
   );
 });
